move modals for record edit to record edit folder
